minor fix for sparse table

+++ code/DataStructure/sparse-table.cpp
@@ -2,7 +2,7 @@ struct sparse{
     int logn;
     vector<vector<int>> f;
     sparse(const vector<int>& a){
-        int n=size(a);
+        int n=(int)size(a);
         logn=__lg(n);
         f=vector(n,vector(logn+1,0));
         for(int i=0;i<n;i++) f[i][0]=a[i];
